CARDS-1375: Add an option to block advancing in a form unless all required fields have been filled out Separated utility functions to a separate utility file
